+++ src/checker/tests/mod.rs
@@ -15,6 +15,7 @@ mod deprecated;
 mod destructuring;
 mod entry_point;
 mod field_init;
+mod final_parent;
 mod function_imports;
 mod generics;
 mod inheritance;